fix(lgtm-fixes): removed inert debug code. Rename file to match class name, fixed references
